empty loss.py
