--- docs/slisemap.loss.md
@@ -1,1 +1,0 @@
-# ::: slisemap.loss
